functions/common/requirements.txt
lambda-decorators
cryptography==2.3
python-jose
pytz